Reorganize according to Francois' style
